feat: add ToggleButton.Row component
